backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-jose[cryptography]==3.3.0

